+++ src/main/java/nayan92/moneytransfer/db/mapper/DbAccountMapper.java
@@ -10,6 +10,6 @@ import java.sql.SQLException;
 public class DbAccountMapper implements RowMapper<DbAccount> {
     @Override
     public DbAccount map(ResultSet rs, StatementContext ctx) throws SQLException {
-        return new DbAccount(rs.getLong("account_id"), rs.getLong("balance"));
+        return new DbAccount(rs.getInt("account_id"), rs.getInt("balance"));
     }
 }
