+++ src/main/resources/db/migration/V1__Initial_schema.sql
@@ -1,5 +1,5 @@
 create table account(
-    account_id bigint auto_increment primary key,
+    account_id int auto_increment primary key,
     balance bigint not null
 --     creation_time timestamp
 );
